+++ cluster-cm/src/main/java/com/sequenceiq/cloudbreak/cm/polling/task/ClouderaManagerSyncApiCommandIdCheckerTask.java
@@ -1,5 +1,6 @@
 package com.sequenceiq.cloudbreak.cm.polling.task;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 import org.slf4j.Logger;
@@ -40,7 +41,7 @@ public class ClouderaManagerSyncApiCommandIdCheckerTask
         boolean result = false;
         ClouderaManagerSyncCommandPollerObject castedObj = cast(pollerObject);
         try {
-            Optional<Integer> commandId = syncApiCommandRetriever.getCommandId(
+            Optional<BigDecimal> commandId = syncApiCommandRetriever.getCommandId(
                     castedObj.getCommandName(), api, castedObj.getStack());
             if (commandId.isPresent() && !commandId.get().equals(pollerObject.getId())) {
                 LOGGER.debug("Found a new latest command ID for {} command: {}", castedObj.getCommandName(), commandId);
