+++ datalake/src/main/java/com/sequenceiq/datalake/controller/sdx/SdxController.java
@@ -344,9 +344,7 @@ public class SdxController implements SdxEndpoint {
     @Override
     @InternalOnly
     public RangerCloudIdentitySyncStatus getRangerCloudIdentitySyncStatus(String envCrn, long commandId) {
-        return rangerCloudIdentityService.getRangerCloudIdentitySyncStatus(
-                envCrn,
-                (int) commandId);
+        return rangerCloudIdentityService.getRangerCloudIdentitySyncStatus(envCrn, commandId);
     }
 
     @Override
